[BUGFIX] Remove console output on startup
When there are no carousel elements, a message was output on the console:

carouFredSel: No element found for ".carousel".

This message was output before the "debug" configuration was evaluated,
causing the console to be filled.

This message is now removed.

This commit also drops trailing whitespace from the file
(automatically done on saving in the editor).